+++ src/query-performance-monitoring/validator/validations.go
@@ -17,12 +17,15 @@ const performanceSchemaQuery = "SHOW GLOBAL VARIABLES LIKE 'performance_schema';
 // Query to get the MySQL version
 const versionQuery = "SELECT VERSION();"
 
+// Stored procedure to enable essential consumers and instruments in the Performance Schema
+const enableEssentialConsumersAndInstrumentsProcedureQuery = "CALL newrelic.enable_essential_consumers_and_instruments();"
+
 // Dynamic error
 var (
 	ErrImproperlyFormattedVersion = errors.New("version string is improperly formatted")
 	ErrPerformanceSchemaDisabled  = errors.New("performance schema is not enabled")
 	ErrNoRowsFound                = errors.New("no rows found")
-	ErrMysqlVersion               = errors.New("only version 8.0+ is supported")
+	ErrMySQLVersion               = errors.New("only version 8.0+ is supported")
 	ErrUnsupportedMySQLVersion    = errors.New("MySQL version is not supported")
 )
 
@@ -57,12 +60,6 @@ func ValidatePreconditions(db utils.DataSource) error {
 	if errEssentialConsumers != nil {
 		log.Warn("Essential consumer check failed: %v", errEssentialConsumers)
 	}
-
-	// Check if essential instruments are enabled
-	errEssentialInstruments := checkEssentialInstruments(db)
-	if errEssentialInstruments != nil {
-		log.Warn("Essential instruments check failed: %v", errEssentialInstruments)
-	}
 	return nil
 }
 
@@ -88,11 +85,13 @@ func isPerformanceSchemaEnabled(db utils.DataSource) (bool, error) {
 }
 
 // checkEssentialStatus executes a query to check if essential items are enabled.
-func checkEssentialStatus(db utils.DataSource, query string, updateSQLTemplate string, errMsgTemplate error) error {
+func checkEssentialStatus(db utils.DataSource, query string, updateSQLTemplate string) (count int, essentialError error) {
+	enabledCount := 0
+
 	// Execute the query
 	rows, err := db.QueryX(query)
 	if err != nil {
-		return fmt.Errorf("failed to check essential status: %w", err)
+		return enabledCount, fmt.Errorf("failed to check essential status: %w", err)
 	}
 	defer func() {
 		if err := rows.Close(); err != nil {
@@ -100,22 +99,36 @@ func checkEssentialStatus(db utils.DataSource, query string, updateSQLTemplate s
 		}
 	}()
 
-	// Check if each essential item is enabled
 	for rows.Next() {
 		var name, enabled string
 		if err := rows.Scan(&name, &enabled); err != nil {
-			return fmt.Errorf("failed to scan row: %w", err)
+			return enabledCount, fmt.Errorf("failed to scan row: %w", err)
 		}
-		if enabled != "YES" {
+		if enabled == "YES" {
+			enabledCount++
+		} else {
 			log.Warn(updateSQLTemplate, name, name)
-			return fmt.Errorf("%w: %s", errMsgTemplate, name)
 		}
 	}
 
 	if err := rows.Err(); err != nil {
-		return fmt.Errorf("query to check essential status failed: %w", err)
+		return enabledCount, fmt.Errorf("query to check essential status failed: %w", err)
 	}
 
+	return enabledCount, nil
+}
+
+/*
+enableEssentialConsumersAndInstrumentsProcedure calls a stored procedure to enable essential consumers and instruments.
+This procedure ensures that the required consumers and instruments in the Performance Schema are enabled.
+*/
+func enableEssentialConsumersAndInstrumentsProcedure(db utils.DataSource) error {
+	_, err := db.QueryX(enableEssentialConsumersAndInstrumentsProcedureQuery)
+	if err != nil {
+		return fmt.Errorf("failed to execute stored procedure to enable essential consumers and instruments: %w", err)
+	}
+
+	log.Debug("Stored procedure executed successfully.")
 	return nil
 }
 
@@ -123,14 +136,14 @@ func checkEssentialStatus(db utils.DataSource, query string, updateSQLTemplate s
 func checkEssentialConsumers(db utils.DataSource) error {
 	query := buildConsumerStatusQuery()
 	updateSQLTemplate := "Essential consumer %s is not enabled. To enable it, run: UPDATE performance_schema.setup_consumers SET ENABLED = 'YES' WHERE NAME = '%s';"
-	return checkEssentialStatus(db, query, updateSQLTemplate, utils.ErrEssentialConsumerNotEnabled)
-}
-
-// checkEssentialInstruments checks if the essential instruments are enabled in the Performance Schema.
-func checkEssentialInstruments(db utils.DataSource) error {
-	query := buildInstrumentQuery()
-	updateSQLTemplate := "Essential instrument %s is not fully enabled. To enable it, run: UPDATE performance_schema.setup_instruments SET ENABLED = 'YES', TIMED = 'YES' WHERE NAME = '%s';"
-	return checkEssentialStatus(db, query, updateSQLTemplate, utils.ErrEssentialInstrumentNotEnabled)
+	count, consumerErr := checkEssentialStatus(db, query, updateSQLTemplate)
+	// If the count of enabled items is less than 3, call the stored procedure
+	if count < constants.EssentialConsumersCount {
+		if err := enableEssentialConsumersAndInstrumentsProcedure(db); err != nil {
+			return fmt.Errorf("failed to enable essential consumers and instruments via stored procedure: %w", err)
+		}
+	}
+	return consumerErr
 }
 
 // logEnablePerformanceSchemaInstructions logs instructions to enable the Performance Schema.
@@ -207,29 +220,6 @@ func buildConsumerStatusQuery() string {
 	query := "SELECT NAME, ENABLED FROM performance_schema.setup_consumers WHERE NAME IN ("
 	query += "'" + strings.Join(consumers, "', '") + "'"
 	query += ");"
-
-	return query
-}
-
-// buildInstrumentQuery constructs a SQL query to check the status of essential instruments
-func buildInstrumentQuery() string {
-	// List of essential instruments to check
-	instruments := []string{
-		"wait/%",
-		"statement/%",
-		"%lock%",
-	}
-
-	// Pre-allocate the slice with the expected length
-	instrumentConditions := make([]string, 0, len(instruments))
-	for _, instrument := range instruments {
-		instrumentConditions = append(instrumentConditions, fmt.Sprintf("NAME LIKE '%s'", instrument))
-	}
-
-	// Build the query to check the status of essential instruments
-	query := "SELECT NAME, ENABLED FROM performance_schema.setup_instruments WHERE "
-	query += strings.Join(instrumentConditions, " OR ")
-	query += ";"
 
 	return query
 }